Error Request changes

+++ products/src/main/java/com/shawoor/products/services/ProductService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -21,11 +22,16 @@ public class ProductService {
     }
 
     public Product getProductById(Long id) {
-        Product dummyProduct = new Product();
-        dummyProduct.setId(0L);
-        dummyProduct.setTitle("No Product Found");
-        dummyProduct.setDescription(" ---- ");
-        return productRepository.findById(id).orElse(dummyProduct);
+//        Product dummyProduct = new Product();
+//        dummyProduct.setId(0L);
+//        dummyProduct.setTitle("No Product Found");
+//        dummyProduct.setDescription(" ---- ");
+
+//        Optional<Product> optionalProduct = productRepository.findById(id);
+//        Product product;
+//        product = optionalProduct.orElse(null);
+//        return product;
+        return productRepository.findById(id).orElse(null);
     }
 
     public List<Product> getProductsByTitle(String searchTerm) {
@@ -65,22 +71,34 @@ public class ProductService {
     }
 
     public Product updateProduct(Long id, Product productDetails) {
-        Product product = productRepository.findById(id).orElseThrow();
-        product.setTitle(productDetails.getTitle());
-        product.setDescription(productDetails.getDescription());
-        product.setPrice(productDetails.getPrice());
-        product.setRating(productDetails.getRating());
+//        Product product = productRepository.
+//        product.setTitle(productDetails.getTitle());
+//        product.setDescription(productDetails.getDescription());
+//        product.setPrice(productDetails.getPrice());
+//        product.setRating(productDetails.getRating());
+        Optional<Product> optionalProduct = productRepository.findById(id);
+        Product product;
+        if (optionalProduct.isPresent()) {
+            return productRepository.save(productDetails);
+        } else {
+            return product = null;
+        }
+//        return productRepository.findById(id).orElse(null);
 
-        return productRepository.save(product);
     }
 
     public Product postProduct(Product product) {
         return productRepository.save(product);
     }
 
-    public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+    public boolean deleteProduct(Long id) {
+        Optional<Product> optionalProduct2 = productRepository.findById(id);
+
+        if (optionalProduct2.isPresent()) {
+            productRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
-
-
 }
